refactor: move local SSD helper to pkg/utils/localssd sub-package
Isolates the dependency-free local SSD sizing logic into its own package
so that pkg/providers/pricing/gcpclient can import it without pulling in
the Kubernetes dependencies that pkg/utils carries (controller-runtime,
karpenter APIs, etc.).

Rename LocalSSDTotalGiB → TotalGiB and DefaultSSDPartitionGiB →
DefaultPartitionGiB to fit the package-qualified call site style
(localssd.TotalGiB, localssd.DefaultPartitionGiB).
